+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   * Added `--stateboard` and `--stateboard-only` options for `start` and `stop`
     commands to start/stop stateboard locally
 - Warning on running `cartridge start` without `cartridge build` before
+- Checking notify socket length on `cartridge start -d`
 
 ### Changed